+++ CHANGELOG.md
@@ -7,8 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Initial Release
+- None
 
 ## [0.1.0] - yyyy-mm-dd
 
-
+- Initial Release